refactor: reorganise existing feedback component folder
